only 1password apt list does not exists then create and install that

+++ brew-env/_linux.sh
@@ -66,17 +66,19 @@ sudo ln -sf /home/linuxbrew/.linuxbrew/share/fonts /usr/local/share/
 sudo fc-cache -fv
 
 # install 1Password GUI
-curl -sS https://downloads.1password.com/linux/keys/1password.asc | \
-	sudo gpg --dearmor --output /usr/share/keyrings/1password-archive-keyring.gpg
-echo 'deb [arch=amd64 signed-by=/usr/share/keyrings/1password-archive-keyring.gpg] https://downloads.1password.com/linux/debian/amd64 stable main' | \
-	sudo tee /etc/apt/sources.list.d/1password.list
-sudo mkdir -p /etc/debsig/policies/AC2D62742012EA22/
-curl -sS https://downloads.1password.com/linux/debian/debsig/1password.pol | \
-	sudo tee /etc/debsig/policies/AC2D62742012EA22/1password.pol
-sudo mkdir -p /usr/share/debsig/keyrings/AC2D62742012EA22
-curl -sS https://downloads.1password.com/linux/keys/1password.asc | \
-	sudo gpg --dearmor --output /usr/share/debsig/keyrings/AC2D62742012EA22/debsig.gpg
-sudo apt update && sudo apt install -y 1password 1password-cli policykit-1-gnome
-sudo groupadd onepassword-cli
-sudo chown root:onepassword-cli "$(which op)" && \
-sudo chmod g+s "$(which op)"
+if [[ ! -f "/etc/apt/sources.list.d/1password.list" ]]; then
+	curl -sS https://downloads.1password.com/linux/keys/1password.asc | \
+		sudo gpg --dearmor --output /usr/share/keyrings/1password-archive-keyring.gpg
+	echo 'deb [arch=amd64 signed-by=/usr/share/keyrings/1password-archive-keyring.gpg] https://downloads.1password.com/linux/debian/amd64 stable main' | \
+		sudo tee /etc/apt/sources.list.d/1password.list
+	sudo mkdir -p /etc/debsig/policies/AC2D62742012EA22/
+	curl -sS https://downloads.1password.com/linux/debian/debsig/1password.pol | \
+		sudo tee /etc/debsig/policies/AC2D62742012EA22/1password.pol
+	sudo mkdir -p /usr/share/debsig/keyrings/AC2D62742012EA22
+	curl -sS https://downloads.1password.com/linux/keys/1password.asc | \
+		sudo gpg --dearmor --output /usr/share/debsig/keyrings/AC2D62742012EA22/debsig.gpg
+	sudo apt update && sudo apt install -y 1password 1password-cli policykit-1-gnome
+	sudo groupadd onepassword-cli
+	sudo chown root:onepassword-cli "$(which op)" && \
+	sudo chmod g+s "$(which op)"
+fi
